requirements.txt
streamlit
PyMuPDF
numpy
faiss-cpu
sentence-transformers
python-dotenv
requests
langchain
youtube-transcript-api
speechrecognition
fpdf
pyttsx3